Add parens to fix precedence in assertion
The corrected check was trivially always true, the performed check
is comparing the 'bool' flag to an integeral (almost always false)
and then taking the fail branch of the ternary, which is a simple
integer that is never zero.

Restored the intended behavior and ran test several times to
confirm the assertion dose not trigger.  Still not clear why we
need an assertion for other than a precondition in the middle
of a function.