Fix bool warning in balm_category.cpp
Warning flagged by Microsoft, that we are promoting a 'bool' to
an 'int' before performing an equality comparison.  In practice
this will not be an issue, as the atomic 'int' in question is
always set by writing a 'bool' value, but better to force a cast
to 'bool' and make the intent clear.

There is a remaining thread-safety concern about the whole idiom
being used in this function in the code protected by this awkward
test, but that is the subject of a separate DRQS.