Remove dependency on std::iterator
std::iterator is deprecated in C++17, and leads to compiler warnings
in appropriate build modes.  The intended replacement is to simply
provide the typedefs needed for 'iterator_traits' directly.  However,
we must retain inheritance from 'std::iterator' when building on
Solaris with the (default) Rogue Wave library, as that non-standard
implementation requires the inheritance in order to dispatch the
standard algorithms correctly.