Fix preprocessor issues for Linux compilers
The stricter preprocessor on sevearl Linux compilers complained about
the attempt to finagle the library features macros in 'bsl_ctsdio.h'.
The fix is to ensure each feature macro is defined to '1' in the
library features header.  Also, ensure that Linux libraries relying
on the platform headers for 'snprintf' are as supported as the UNIX
platforms when compiling as C++03.