Finish MSVC C++17 enablement
This patch concludes efforts to support building on MSVC with C++17
language support turned on, but does not yet attempt to expose the
C++17 library features through namespace 'bsl'.

The bulk of the patch is replacing calls to the deprecated (and now 'random_shuffle'
removed) 'random_shuffle' with calls to 'shuffle', which requires
the user to supply a random bit engine as the source of randomness.

Other fixes cover the removal of 'bind2nd', used in exactly one
header and conditionally replaced with use of a lambda expression
to avoid worrying about package levelization.  The language detection
of lambda support is local to this file, as there was no call to
promote a feature macro up to 'bsls_compilerfeatures' for a single
usage, although we should think about properly supporting a lambda
detection macro as a follow-up patch.

Similarly, there is one test driver that actually makes use of
incrementing 'bool' to ensure that there are no odd platform bugs
when hashing a 'bool' value (i.e., increment does not change the
hidden value when it is 'true').  A feature macro for this is to
be actively avoided, lest we accidentally inspire C++14 and earlier
consumers to actually use the feature!

Finally, there were some doc issues to address related to these
changes, and minimal test coverage noting the addition of new
library headers under the various 'BASELINE' macros.