Correct the contract on printAttribute
The bslim_printer offers a 'printAttribute' method taking a 'name'
argument.  The type of the argument was changed from 'const char*' to
'const bslstl::StringRef&' at the same time as an assert on 'name != 0'
was removed from the implementation.  The contract despite the assertion
assigned meaning to the case 'name == 0'.  This case can no longer occur
with the new type, so the wording is removed.

Signed-off-by: Frank Birbacher <fbirbacher@bloomberg.net>