Doc fix support iterators (#2219)
* Updated docs for iterator base types

Attempt to better clarify the constructor that is sometimes a
copy constructor, and sometimes an overload. per DRQS 123560143.
Follow standard convention that 'value_type' for an iterator is
never cv-qualfied, but rather, the 'pointer' and 'reference'
types hold the cv-qualifier.

* Update stl proto-iterators to using bsltestutil

Update the test drivers for 'bslstl_forwarditerator',
'bslstl_bidirectionaliterator' and 'bslstl_randomaccessiterator' to
using the standard test utility and test macros.  Looked to add
missing test coverage for the case of copy-assignment for a non-const
iterator to a const-iterator.  Turns out this is already tested, and
the functionality is (intentionally) present through creation of a
temporary using the implicit conversion constructor.