Clean up bslstl_hash test driver (#2221)
The test driver for bslstl_hash was carrying baggage from a long
passed transition of how C-string values were hashed.  As part of
cleaning out the expired comments (that indicate they were supposed
to be removed at least 3 years ago, likely longer) update the test
driver with a round of conformity tweaks for the latest bde_verify,
and apply matching fixes to the test driver in the component header
where relevant.