Fix 'bslstl_pair.h' warnings on Clang (#2247)
Clang has two warnings.  First first is annoying and must be worked
around with #pragmas, as different standard library implementations
define tuple API template specializations using different 'class'
and 'struct' tags, often in the same library so that we cannot use
a consistent tag ourselves, even with platform detection.

The second warning is a much simpler function declaring a named
parameter that is not used in the function.  This occurs when we
pass an unused allocator to functions with a tag-type indicating
that the allocator will be unused.  Most of these uses already
removed the parameter name on the function defintiion; this patch
catches the last few cases, which Clang likes to warn about.