First cut at 'bdlb_nullopt'
Basic implementation of the component, with necessary
update to 'bsls_keyword.h'.  Currently lacking a usage
example, and the testing for the new macro added to
'bsls_keyword.h'.  The keyword test driver was updated
to at least reflect which of the current macros are
not tested, and component and testing checklist were
alpha-sorted for the declaration of macros, to aid
future QA audits.