Simplifying bslmf_voidtype for better use with C++11 (#2280)
* Simplifying bslmf_voidtype for better use with C++11

This patch provides a proper 'bsl::void_t' that is a simplfied alias
template for C++11 using variadic macros, to simplify the number of
template instantiations when using the SFINAE technique enabled by
this component.  In addition, the macro 'BSLMF_VOIDTYPE' will expand
into the correct idiomatic use for both C++11 and C++03, simplifying
usage.  Existing usage of this trait in 'bslmf' is updated accordingly.

Then, 'bslmf_detectnestedtrait' is greatly simplified using this
trait, as a demonstration of how we might radically simplify some
traits, reducing the number of template instantiations required for
common use cases.

Finally, the 'is_function' trait is simplified for IBM using
'BSLMF_VOIDTYPE', and this implementation is adopted by the Sun CC
compiler to avoid warnings that cannot be silenced with a simple pragma.

A follow-up change-set will apply the revised trait across the rest of
'bslmf' where it proves beneficial - notably in reducing the overhead
of many template instantiations for 'is_convertible'.

* Apply review feedback

Mostly documentation fixes. plus a correction for compiler version
detection for a workaround in the 'VoidType' test driver.

* Further editorial review changes