Review use of DetectNestedTraits for consistency
Review all 'bsl' uses of 'bslmf::DetectNestedTrait' to ensure that
trait specialization by the user produces a consistent (and expected)
result for all cv-qualfied types.  Update test drivers where test
coverage was lacking.

Note that the 'DetectNestedTrait' idiom will fail to compile for
incomplete class types, and this should be consistently documented.

Also note that while Sun CC 12.3 and earlier DO parse abominable function
types, they false match them in partial specializations as cv-qualified,
so any attempt to delegate a cv-partial specialization to a primary trait
will be infinitely recursive on these compilers.  As such, we disable
testing of abominable function types on these old platforms.  We could
provide workarounds for the affected components so that the tests would
pass, but it is deemed impractical to add extra template instantiations
on these traits in order to support obscure corner cases that will loudly
fail to compile, rather than silently do the wrong thing.  This affects
any trait that intends to support specialization of only the primary
template to handle all cv-qualified variations of a type as well, including
any trait that also supports nested trait declarations, or additional type
traits implemented in terms of such traits.

Finally addressing the cv-qualifier issues pushed through some pragmas to
silence warnings in test drivers on Solaris, so the general task of
cleansing Solaris of false-positive warnings by use of pragmas was also
completed (for bsl).