Resolve 'constexpr' and 'noexcept' for 'bsl::array'
The final check-in for 'bsl::array' managed to miss all the 'constexpr' and 'noexcept'
markup required by the standard.  This patch should be compatible with both C++11 and
C++14 standards, using feature-detection macros as appropriate.

Test driver enhanced to validate the correct use of 'constexpr', but validating
'noexcept' left as a follow-up exercise in order to get this patch into the hands
of users as soon as possible.