Clean up basic bde_verify issues in bslstl test drivers (#2322)
* Clean up basic bde_verify issues in bslstl test drivers

This pass is picking up simple issues such as correct banners
and ASSERT macro definitions.

* Fix for review feedback

The hash table test driver fails to build when switching from namespace
'bsl' to namespace 'std' for the <iterator> related parts.  It is not
clear that the bde_verify warnings were that helpful in this case, so
revert that subset of changes.