Native_std forward list (#2210)
* Initial support for 'native_std::forward_list'

This patch pulls support for the native stl 'forward_list'
container into namespace 'bsl' using the bsl allocator to
properly support other allocator-aware types.  It comes
with only minimal testing.

Note that this initial implementation does NOT support
types that simply specialize the 'UsesBslmaAllocator'
trait, as it relies on deducing allocator support through
the native STL 'allocator_traits', which are looking at
specializations of 'native_std::uses_allocator'.  This
trait is satisfied by types that have a nested typedef
for 'allocator_type' that is a specialization of
'bsl::allocator<T>'.

* Fix allocation count for MSVC

MSVC containers allocate an extra sentinel node
in their list containers, so allow for an extra
allocation when building with MSVC libraries.

* Resolve review feedback on documenting test driver