Remove external header guards from bsl+stdhdrs (#2335)
It looks like the previous scripts run over 'bsl' failed to pick
up on the headers (with non-standard names) in the 'bsl+stdhdrs'
directroy, and also their unusual indenting style.

This patch manually reviews and fixes up each file, along with
applying a consistent (simpler) indent for the preprocessor that
is easier to read.  Generally, indenting on #if control-flow, but
preserving the vertical sense of a repeated macro name in cases
such as ifdef/undef etc.