Avoid a MetaInt temp when dispatching RangeCompare implementations
The internal dispatch logic for switching to optimized algorithms in RangeUtil
overloads on 'true_type' and 'false_type', but creates a 'MetaInt<0 or 1>' to
dispatch and then slices.  This is a legacy of the pre-'bsl' days when traits
used a 'Type' rather than 'type' typedef for the resuilt type of trait queries,
and we retain the old name as an alias to the old type.

Surprisingly, this conversion showed up as consuming around 5% of time in a
benchmark for code making heavy use of RangeCompare.

The fix is to simply use the lower-case 'type' instead, avoiding the extra
temporary.