Update platform report in bsls_platform.t.cpp
Given the number of macros that are tested, it is extremely difficult to see
which macros are defined for a given platform among the sea of undefined macros.
Revise the test report to collect all the undefined macros to separately report
them at the end.  Then revise the classification of the macros that are reported.

This patch removes quite a few duplicate checks from the report function, which
appeared to have two long overlapping list of macros to test and report, with
each list having an extra 5-10% macros not present in the other.