Correct is_nothrow_move_constructible trait (#2409)
* Correct is_nothrow_move_constructible trait

The trait 'bslmf::is_nothrow_move_constructible' is designed to be a user
extensible approximation of the C++11 trait.  However, it behaves quite
differently on even built-in types, and generally gives the wrong answer
for cv-qualified class types when we opt-in.

This patch updates the test driver so that a C++11-enabled build can
observe whether our customized trait matches the behavior of the built-in
trait for native types, and whether any divergence in behavior for types
using BDE trait customization is to be expected, both in 03 and in 11.

* Refactor bslmf tests for bad Oracle on Microsoft

MSVC 2013/5 have bad results for 'is_nothrow_move_constructible' and
rvalue-references to arrays.  We refactor our test macros so that the
testing allows for a broken Oracle, rather than fixing our own
implementation to have bad results.

* Fix additional wrinkles in traits testing for MSVC 2013

* Clean-up, and fix for broken macro definition in C++11 builds

* Resolve review feedback on 'bslmf_isnothrowmoveconstructible'