Simplifying 'bslmf_voidtype' for more efficient use with C++11
This patch provides a proper 'bsl::void_t' that is a simplified alias
template for C++11 using variadic macros, to reduce the number of
template instantiations when using the SFINAE technique enabled by
this component.  In addition, the macro 'BSLMF_VOIDTYPE',
'BSLMF_VOIDTYPE2', and 'BSLMF_VOIDTYPES' will expand into the correct
idiomatic code for both C++11 and C++03, simplifying usage.  Existing
usage of this trait in the BDE repository is updated accordingly.

The macros taking a reduced number of arguments are an attempt to provide
more efficient template on legacy C++03 compilers, as by far the most
common use case is a single argument.  The two-argument form is provided
in anticipation of further clean-up patches to reduce the number of
template instantiations throughout 'bslmf'.