Change null+nil XML encoding to absence and add bdlat testing apparatus
**NOTE** This commit changes an encoding created by `balxml::Encoder` in
a way that is backwards-compatible with the existing `balxml::Decoder`.
However, third party decoders may decode the new encoding differently
than the previous one.  Even though the pre-commit encoding behavior is
technically not valid against the corresponding schema, the change may
still cause issues for third-party decoders, and we should NLRT this
change.

This commit addresses {DRQS 139782067}, which identifies an issue in the
way that `balxml::Encoder` encodes attributes of `bdlat` types that have
the `Sequence` `bdlat` type category and have the
`bdlat_FormattingOptions::e_NILLABLE` formatting mode option enabled.

Prior to this commit, such fields would be encoded as a self-closing tag
having no attributes `<LikeThisForExample/>`.  However, this is not a
valid encoding of a nullable and nillable field according to XSD.  The
only two valid encodings of such a field are to omit it in the XML, or
to encode it as a self-closing tag with the attribute `xsi:nil="true"`.
(Technically, the two encodings represent two different values, the null
value and the nil value, respectively, but these are frequently treated
as meaning the same thing.) Since the encoder can only encode such a
field in the latter form if the `objectNamespace` option is non-empty
and the `outputXSIAlias` option is not `false`, the default encoding
must omit such fields.

This commit changes the encoder to omit such fields unless the
`objectNamespace` and `outputXSIAlias` options meet the above criteria,
in which case these fields are encoded as self-closing tags with
the attribute `xsi:nil="true"`.

Prior to this commit, the test drivers for both the `balxml_encoder` and
`balxml_decoder` components had very poor coverage of the `bdlat` type
categories.

This commit introduces internal testing infrastructure to thoroughly
exercise the space of `bdlat` types in the encoder and decoder.

Deployment Risks:
This encoding change is backwards-compatible with the existing decoder,
but may not be backwards-compatible with the expectations of third
party decoders, or hand-written XML processing tools that expect
the prior encoding.  Such systems may begin to fail to decode XML
encoded using this encoder once it is deployed.  However, systems
like BAS and other infrastructure using `balxml::Decoder` will
continue to decode such encodings to the same value.

Test Plan:
Test drivers added.  All tests pass.

Ticket:
{DRQS 139782067}