ManagedPtr documentation fix re: default allocator.
The `ManagedPtr` documentation, in both the component documentation and
in the documentation for the non-factory, non-deleter constructor
incorrectly states that the managed object will be destroyed using the
currently intalled default allocator.  The code and tests actually
intend and do destroy the object using `delete ptr` in this case, so
the documentation is updated herein to reflect that.

DRQS 148082770