Applications of improved 'BSLMF_VOIDTYPE' facility
Now that we can easily select qualifying partial specializations with
'BSLMF_VOIDTYPE', deploy this macro/trait to reduce the number of template
instantiations in a few common cases, and to simplify the implementation
of compile-time reflection machinery, requiring fewer compiler workarounds.

The most notable improvement is to the implementation of nested traits
detection.