Fix error propagation bug in 'balxml_decoder'
Prior to this commit, there was a bug in `balxml::Decoder::decode` that
could cause an error code returned by an object being decoded-into to
be ignored.

This commit modifies `balxml::Decoder::decode` to recognize and
propagate that error code.

Test Plan:
Tests added and passed. Confirmed prior code fails tests.

Deployment Risks:
In theory, this change can cause the decoder to start returning non-zero
error statuses given input for which it previously returned zero.
However, for all input where that would be the case, the decoded value
would be partially formed or corrupt.  In addition, the input that would
have triggered this behavior in the past is difficult to construct, and
likely rare in the wild. (For example, the codec had existed for many
years before the ticket that identified this issue was filed.)