balxml_encoder/decoder: Ensure 'loggedMessages' clear on 'decode'
This commit addresses a bug in the implementations of
'baljsn::Encoder::encode' and 'baljsn::Decoder::decode'.

The contracts of each function respectively state that they reset the
'loggedMessages' buffer of the corresponding object.  However, prior
to this commit, each operation would fail to clear the 'loggedMessages'
buffer in some cases.  In particular, some checks were performed early
enough in the en/de-coding process that, if they failed, the failure
would log a message and return to the caller before the code to clear
the logged messages was invoked.

This commit changes 'baljsn::Encoder::encode' and
'baljsn::Decoder::decode' to unconditionally clear their respective
'loggedMessages' on each invocation, no matter how early the operation
fails.

Test Plan:
Tests added.

Deployment Risks:
The 'loggedMessages' attributes of the codecs provide unspecified,
human-readable error messages.  Changing their contents, or when they
are cleared, is within the existing contracts of the respective
functions.