DRQS 112643766: Remove UB in 'balber_beruniversaltagnumber'
This changeset removes undefined behavior that was present in
the implementation of `balber::BerUniversalTagNumber::select`.

Prior to this changeset, the implementation relied on driving overload
resolution by creating a reference to the type used to select an
overload by dereferencing the null pointer.  It did this in 2 places.

These changes modify the implementation of the functions that
dereferenced the null pointer so that they no longer need to do so.  The
behavior of `balber::BerUniversalTagNumber` is left unchanged.

Test Plan:
Verified new test passes with old and new implementation.

Tickets:
{DRQS 112643766 <GO>}

Risks:
Although this changeset is designed to leave the beahvior of
`balber::BerUniversalTagNumber::select` unmodified, the function is a
template, and it is possible that the results returned by the new
implementation are different than that of the old for some type not
verified in the test driver.  However, the test driver verifies all
types that are in-contract.