Reduce clang-tidy and coverity warnings when using 'bslmt_once'
This commit is based on the work provided in
https://bbgithub.dev.bloomberg.com/bde/bde/pull/2520

Prior to this commit, `clang-tidy` would report lines expanded from
`BSLMT_ONCE_DO` macro expansions in warnings related to it not being
able to determine whether the "once block" would be evaluated.

Additionally, Coverity would report that execution could continue past
the `BSLS_ASSERT` on line 43 of `bslmt_once.cpp`.  This is technically a
bug in `bslmt_once`, as a build with `BSLS_ASSERT` disabled, or with a
continuing handler, could cause execution to continue past that point.

This commit adds a `/* NOLINT */` comment to the appropriate position in
the definition of `BSLMT_ONCE_DO` to exclude the expanded macro contents
from appearing in the trace of any `clang-tidy` warnings.
Unfortunately, this does not suppress such warnings entirely.

It also addresses the issue identified by Coverity.

Tickets:
{DRQS 149823688}

Test Plan:
Existing tests pass.

Deployment Risks:
Minimal, this change is almost entirely cosmetic.  It is possible that
builds with `BSLS_ASSERT` turned off could have produced executables where
execution did continue past line 43 of `bslmt_once.cpp`, however, that
execution path is library undefined behavior and the program would
be out of contract anyway.  This commit does change the behavior such
that the program will always terminate if execution passes reaches that
line.