DRQS 151245190: Fix Failing baljsn_encoder Tests on Sun
This commit fixes the test failures for optimized builds
of `baljsn_encoder` on Sun.

Prior to this commit, a table in a table-based test case did not
properly ensure some data was zero-initialized where zero-initialization
was required.

This commit modifies the definition of the table structure to ensure
that it is statically zero-initialized.  It also applies this
modification to a table in the test driver of `baljsn_decoder`, which
asymptomatically had the same problem.

Test Plan:
Broken tests now pass.

Tickets:
{DRQS 151245190 <GO>}

Deployment Risks:
None, this is testing-only change.