Immediate patch to unblock the nightlies build
This patch simply fixes the immediate build failures for strict
C++03 builds that do not support reference collapsing, which are
now showing up on the nightly builds.

A follow-up ticket will audit containers and confirm that
reference types as elements have all needed test coverage, and
perhaps excessive support that is no longer demanded for C++20
is relaxed, potentially simplifying some of the more excessive
test drivers such as for bslstl_hashtable.