Fix library detection used for Oracle CC workarounds
There are several remaining cases of using the 'BSLS_PLATFORM_OS_SOLARIS'
macro as a proxy for detecting the broken Rogue Wave standard library
used by default on Solaris.  Now that we have the library features
component providing proper library detection macros, and are testing on
Solaris with a wider variety of compilers, we should update the workaround
detections accordingly.