Yet another attempt to fix MSVC is_trivially_copyable test results
After several attempts to clean the build, admit default that the
MSVC native implementation of 'is_trivially_copyable' has too many
broken edge cases to be a reliable oracle after all.  Remove it
from the set of trusted oracles, and the tests should pass, while
not giving up on platforms where the native oracle is reliable.