Resolve gcc9 issues with bslmt_testutil.t.cpp
gcc9 has changed the way it computes __LINE__ number when a macro
invocation spans multiple lines, specifically in the case of:
   BSLMT_TESTUIL_LOOPX_ASSERT(I, J, K, L, M, N,   // gcc 9+
                              some predicate);    // gcc 8-

The line reported by this macro expansion changes from reporting
the second line in gcc8 and earlier (plus all other Bloomberg
supported compilers) to reporting the first line in gcc9 and later.
This change appears intentional, and following recommended best
practices for the upcoming C standard that specifically tries to
address this situation.

The issue that arises is comparing error-reporting strings in the
tests for the ASSERT family of macros.  The solution, as only the
test driver is impacted, is to rename variables so that no line has
to wrap to conform to the BDE coding rules.

The second minor issues addressed is warnings from a couple of
member functions defined in the test machinary, but never used.
I opted to remove them entirely, on the understanding that if we
ever need a more fully featured 'OutputRedirector', there is a
perfectly good one sitting in the 'bsls_outputredictor' component,
complete with test driver.