DRQS 153917503: Provide Extended-Binary BER Date/Time Format (#2790)
* DRQS 139634693 Start Feature - Monday, November 18 2019

* DRQS 139634693 WIP - Thursday, November 21 2019

* DRQS 139634693 WIP - Friday, November 22 2019

* DRQS 139634693 WIP - Monday, November 25 2019

* DRQS 139634693 WIP - Tuesday, November 26 2019

* DRQS 139634693 WIP - Monday, December 1 2019

* DRQS 139634693 WIP - Tuesday, December 3 2019

* DRQS 139634693 Refactoring Complete - Wednesday, December 4 2019

* DRQS 139634693 WIP - Wednesday, December 18 2019

* WIP - Wednesday, December 18 2019

* WIP - Thursday, December 19 2019

* WIP - Friday, December 20 2019

* DRQS 139634693 WIP - Tuesday, January 14 2020

* DRQS 139634693 - Refactor 'balber_berutil' for future extension

This pull request heavily refactors `balber_berutil` without changing
any behavior of the component.

The encoding and decoding operations for each "Simple" type are
reorganized into individual component-private utility structs.

Additionally, the textual and binary formats for the date and time types
are enumerated, with operations for selecting the format to encode, and
detecting the format from an encoded value.

The purpose of the above is to permit a planned future commit to provide
new formats for encoding date and time types.  This pull request
restructures `balber_berutil` to be easily amenable to extension.

Test Plan:
All tests pass

Deployment Risks:
Despite passing recently introduced tests, it's possible that the
refactoring changes a corner-case behavior.  In the worst case, this
could lead to corrupt encodings, or incorrectly decoded values, of
"Simple" types.  This could lead to creating corrupt BAS messages when
encoding, or misinterpreting the value of messages when decoding.

The cost of potentionally introducing such a defect is high.

Tickets:
{DRQS 139634693}
{DRQS 153917478}

* WIP - Friday, January 24 2020

* WIP - Tuesday, January 28 2020

* WIP - Wednesday, January 29 2020

* WIP - Wednesday, January 29 2020

* WIP - Friday, January 31 2020

* WIP - Monday, February 3 2020

* WIP - Monday, February 3 2020

* DRQS 139634693 WIP - Monday, March 3 2020

* DRQS 139634693 WIP - Tuesday, March 3 2020

* DRQS 139634693 WIP - Tuesday, March 3 2020

* DRQS 139634693 WIP - Tuesday, March 3 2020

* DRQS 139634693 Fix Matrix Build Failures - Thursday, March 12 2020

* DRQS 139634693 Fix Matrix Build Warnings - Wednesday, March 18 2020

* DRQS 153917503: Implement Part 3

* DRQS 153917503: Begin polishing rebase

* DRQS 153917503: Continued polishing rebase

* DRQS 153917503: Continued polishing rebase

* DRQS 153917503: Refactored parameter order for 'get' functions

* DRQS 153917503: Begin refactoring date and time header type

* DRQS 153917503: Continued refactoring

* Finished documenting changes

* Finished test driver documentation.

* WIP

* Addressed review feedback

* Fix integer literal truncation in assert by widening literal