Resolve false-positive warnings on Solaris test drivers
This applies the relevant warning pragma to silence the noisiest
false-positive warnings on Solaris that cannot otherwise be coded
around - mostly as test drivers poke into all cases.  In many
cases there is a pre-existing pragma to silence similar warnings
on Windows and Linux.

The most annoying case is not being able to properly silence the
warning in the 'is_function' type trait, as the warning is issued
at the point of template instantiation, so guarding the template
code has no effect.