Fix build after merging PR 2789 (#2850)
The pull request 2789 introduced a build failure because it was working
against an outdated version of `bdlde_utf8util`.  This pull request
fixes all uses of outdated names, and makes 1 change in a test driver
where the behavior of `bdlde_utf8util` on master had an enhancement not
present in the version from the pull request.