Refactor `baljsn_encoder` test driver (#2882)
This change set provides 2 modifications to the `baljsn_encoder` test
driver that change no behavior of the component.  I created this change
set because I am currently working on a bug fix for `baljsn_encoder`,
and all of the generated data it contains made my static analysis
tools and editor too slow to effectively use.

First, it removes some 30,000 lines of generated JSON and XML data from
the start of the test driver.  This change set moves the data to a
testing-only component in the `s_baltst` package.

Second, it refactors some of the hand-written code in the test driver to
consolidate testing machinery.