DRQS 153564520: Make `baljsn_encoder` fail on stream buffer error (#2883)
This change set fixes a bug where all overloads of
`baljsn::Encoder::encode`, which encode the JSON representation of a
supplied value to the output sequence of a `bsl::streambuf`, would
ignore return values that signal error conditions when they invoke
output operations on the stream buffer.

To fix this bug, this change set modifies `baljsn::Encoder::encode` to
return a non-zero value and log an error message to the encoder's error
log attribute if the stream buffer indicates an error condition after
any output operation.

This change set modifies the observable behavior of the component.

Calls to `baljsn::Encoder::encode` that completed with truncated output
and a 0 return value will now return a non-zero value.

This has a slim chance of breaking a system if it was previously
successfully encoding truncated output.  However, this behavior is
likely to be a bug, anyway.