Fix sljit platform detection with xlc on AIX
The XL C compiler's _ARCH_* macros are based on the -qarch setting and
do not necessarily imply the object mode. In particular the _ARCH_PPC64
macro only indicates that the compiler is targeted to run on a Power
processor with 64-bit support. The __64BIT__ macro is needed to detect
that the compiler is actually in 64bit mode.