bdlcc_skiplist: Fix Clang C++03 Build (#3304)
This change set modifies the expression used in a `sizeof` to calculate
the size of a `SkipList_Node::Ptrs` object.  Prior to this change set,
the code used illegal member access syntax that is an extension
permitted by most modern compilers.  Unfortunately, Clang does not
permit this syntax in C++03 mode.

The expression in the `sizeof` is changed to an equivalent expression in
standard C++.