Removed BSLMF_ASSERT that won't compile on C++03. (#3438)
Since compile time floating point types and precision or even base is
not guaranteed to be the same as run time even on C++11 (and later)
where they are allowed, I have removed the `BSLMF_ASSERT`.  If that
assert would fail, nothing would work in Ryu anyway.

Also added a missing header.