Cleanup for C++20 testing on a Mac M1
This patch supplies additional cleanups that were independently
developed testing C++20 on an M1 mac.  After rebasing onto the
changes already landed in `master`, the main effect is to adopt
an additional library feature macro, in the style of existing
feature macros, rather than rely on inline tests for the C++
language version.  Also added platform support for the Arm 9 CPU,
and fixed up a typo or two.