ball: Use `bsl::string_view` for categoryName/fileName/message (where possible) (#5442)
Currently, the `categoryName`/`fileName`/`message` parameters used in
the `ball` logging interface are passed using `const char*`.  This
requires that these strings be null-terminated.

This commit changes these parameters (where possible) to take
`bsl::string_view`.  This is a backwards-compatible change because there
is an implicit conversion from `const char*` to `bsl::string_view`.
Note that this *moves* the computation of the string length to the
interface rather than the implementation, however it does not *add* any
new overhead.

The motivation for making this change is that we currently integrate an
external library's logging with `ball`, and the external library is
making a switch to use `std::string_view` (see
https://chromium-review.googlesource.com/c/chromium/src/+/6416567 for
reference).  Without this corresponding change in `ball`, it would
require that we construct a temporary `bsl::string` in order to pass the
(null-terminated) `const char*` to `ball`.

Note that the `categoryName` in the `ball_log` and `ball_loggermanager`
component still use `const char*`.  The reason for this is because the
`LoggerManagerConfiguration::CategoryNameFilterCallback` requires using
`const char*`.  However, this will not affect our integration with the
external library, because we do not use this name filter feature in our
integration (we will construct `ball::Category` directly).