Use the correct mecro to enabled warning suppression on Clang (#5515)
The macro `BSLS_PLATFORM_HAS_PRAGMA_GCC_DIAGNOSTIC` should be used
to detect whether the active compiler can use
  `pragma GCC push diagnostic`
to disable a warning message, rather than just detecting for gcc.

This patch applies the correct macro to the few places that used the
smaller test condition for deprecation warnings.  We specifically
avoided checking for all warnings being disabled in this way as some
of those checks could be specifically for warnings in gcc that are
not relevant to the Clang compiler.