Establish compiler features macros that indicate full support for a language standard. (#5552)
* Add feature test macros to indicate full support for a standard

Add several new macros, `BSLS_COMPILERFEATURES_FULL_CPP11` etc., that
indicate a reliable implementation of the core language specified by
that iteration of the C++ Standard.  These pacakging macros are tested
exensively in the component .cpp file, not the test driver, to ensure
that a misconfigured build of BDE will catch such errors, without
placing a burden on every include of the header.  A minimal consistency
check *is* applied in the component header though, to assert that if
a full feature macro is active, then the compiler is currently
configured to support at least that iteration of the standard.

* Add workarounds and tests for Clang not defining feature macros

The Clang compiler had a bug for many releases where it did not
define the standard feature test macros for two features it has
implemented since Clang 7.0 (or earlier).  Remove tests for
those macros for just the affected Clang releases, but add tests
to the regular test driver to ensure that the features work,
even in the absence of the standard test macro.

* Improved documentation suggestion

Review suggests a more reader-friendly way to describe
when a user can rely on the full language macros.

* Note that Clang takes an extra decade for tiny feature

Runtime testing shows that Clang does not implement the C++14 feature
'sized deallocation' until a decade after C++14 support was complete.
We choose to note this missing feature in the full language macro
documentation rather than defer C++14 and C++17 support by a decade.