Fix false positive static analysis warning on memory overwite (#5560)
There are several places throughout our code where we use `if` with a
condition that is fully determined at compile time, such as testing the
values of type traits.  We expect the compiler to optimize the branch
away, but that this entirely up to the optimizer.

When viewed through the lens of a static analyzer, both branches will
be considered, so in the case of `bslalg_scalarprimites::swap` called
for objects of two diffenent sizes, although there is a clear check
against calling the (reported) bad branch, the static analyzer cannot
see it.

The fix is to use C++17 `if constexpr` to force the check to compile
time.  As all of our analyzers reporting this warning are run in
C++20 mode the feature will always be available.  By using the
`BSLS_KEYWORD_CONSTEXPR_CPP17` macro, we will retain the regular
`if` statement for earlier dialects, where we are not running any
tools that would report the false positive.