Make hash table anchors trivially copyable (#5565)
While the `bslalg::HashTableAnchor` class is marked as
bitwise copyable, and hence bitwise movable, it is not in
fact trivially copyable as we provide definitions for both
the copy constructor and copy-assignment operator that do
exactly the same thing as the implicitly declared functions
would, but non-trivially.  Removing the declaration and
user-supplied defintion for these functions satisfieds the
true language requirements for a trivially copyable type
and so silences a few warnings in builds where compilers
are more aggressive about warning some  undefined behaviors.